requirements.txt
fireworks-ai>=0.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
